// src/app/domain/enums/SystemEnums.ts
import { EnumMetadata } from "./EnumMatadata";


export interface SystemEnums {
  rolesUsuario: EnumMetadata[];
  tipoActividad: EnumMetadata[];
  alcance: EnumMetadata[];
  categoriaProductoTransportado: EnumMetadata[];
  medioDeTransporteLogistica: EnumMetadata[];
  tipoTransportePublico: EnumMetadata[];
  tipoVehiculo: EnumMetadata[];
  periodicidad: EnumMetadata[];
  tipoFactorEmision: EnumMetadata[];
  tipoDocumento: EnumMetadata[];
  tipoOrg: EnumMetadata[];
  estadoInvitacion: EnumMetadata[];
  tipoMedioDeTransporte: EnumMetadata[];
}